% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatLoad.R
\name{heatLoad}
\alias{heatLoad}
\title{Calculate allowable effluent temperature discharge from heat loading to a recieving water for wasteload analysis
Follow Utah Rivers Model procedure (roughly 2004) with some modifications
Function uses seasonal receiving and effluent discharge and temperature for heat load calculation
Funtion uses Aquatic Life Classification for stream standard max temperature and max temperature change
Function calculates
Function returns seasonal effluent temperature limits table}
\usage{
heatLoad(
  critQ_cfs,
  effluentQ_cfs,
  temp_receiving,
  temp_effluent,
  aquat_life_use
)
}
\arguments{
\item{critQ_cfs}{Upstream seasonal 7Q10 critical flow in CFS}

\item{effluentQ_cfs}{Seasonal average effluent discharge in CFS}

\item{temp_receiving}{Receiving water temperature C}

\item{temp_effluent}{Effluent discharge water temperature C}

\item{aquat_life_use}{\itemize{
\item aquatic life use. One of 3A, 3B, or 3C
}}
}
\value{
TBD, table, list, etc

Returns a named list of calculated results.
}
\description{
Calculate allowable effluent temperature discharge from heat loading to a recieving water for wasteload analysis
Follow Utah Rivers Model procedure (roughly 2004) with some modifications
Function uses seasonal receiving and effluent discharge and temperature for heat load calculation
Funtion uses Aquatic Life Classification for stream standard max temperature and max temperature change
Function calculates
Function returns seasonal effluent temperature limits table
}
\examples{
result = heatLoad(critQ_cfs=32.0, effluentQ_cfs=7.4326,temp_receiving=13.93, temp_effluent=5.00, aquat_life_use="3B")
}
