% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixZone.R
\name{mixZone}
\alias{mixZone}
\title{Calculate mixing zone statistics for wasteloads
Follow STREAMIX I procedure EPA Region 8, Denver, CO, Water Mgmt Div, Dec 1994
Function returns table with distance from discharge, travel time and plume width
Function returns plume width at 2500 feet distance from discharge and 15 minutes of travel time, expressed as \% of stream width}
\usage{
mixZone(
  critQ_cfs,
  effluentQ_cfs,
  shore_dist_ft,
  width_ft,
  depth_ft,
  slope,
  mix_coeff = 0.6,
  Q_location = "side"
)
}
\arguments{
\item{critQ_cfs}{Upstream seasonal 7Q10 critical flow in CFS}

\item{effluentQ_cfs}{Seasonal average effluent discharge in CFS}

\item{shore_dist_ft}{Discharge distance from shore in feet.}

\item{width_ft}{Stream channel width in feet}

\item{depth_ft}{Stream channel depth in feet}

\item{slope}{Stream channel slope, unitless (typically ft/ft)}

\item{mix_coeff}{Mixing coefficient estimate. Default is 0.6.}

\item{Q_location}{Discharge channel location. One of "side" (default) or "center".}
}
\value{
TBD, table, list, etc
}
\description{
Calculate mixing zone statistics for wasteloads
Follow STREAMIX I procedure EPA Region 8, Denver, CO, Water Mgmt Div, Dec 1994
Function returns table with distance from discharge, travel time and plume width
Function returns plume width at 2500 feet distance from discharge and 15 minutes of travel time, expressed as \% of stream width
}
\examples{
mixing_zone_stats=mixZone(critQ_cfs=736, effluentQ_cfs=2.3, shore_dist_ft=15, width_ft=300, depth_ft=1.8, slope=0.0010, mix_coeff=0.6, Q_location="side")
}
