% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nQy.R
\name{nQy}
\alias{nQy}
\title{Generate low flow statistics (nQy) for wasteloads}
\usage{
nQy(
  data,
  n = 7,
  y = 10,
  date_col = "Date",
  q_col = "discharge_cfs",
  min_days = 328,
  plot_fit = TRUE
)
}
\arguments{
\item{data}{Input data. A dataframe columns of daily discharge values and dates.}

\item{n}{Numeric. Days for rolling time period averages. Default 7.}

\item{y}{Numeric. Recurrence interval in years}

\item{date_col}{Column name containing date values. Must be in standard "YYYY-MM-DD" date format.}

\item{q_col}{Column name containing discharge values. Any appropriate and uniform discharge units are OK. The result will be in the same units as the input.}

\item{min_days}{Minimum number of days required to include a year in the n-day annual minima. Recommend 328 for annual statistics and 82 for seasonal (~10\%).}
}
\value{
Returns a list of results including the calculated nQy statistics (from fitted probability and 1/y percentile), a data frame of annual n-day flow minima and distribution information, and a plot to examine the fit.
}
\description{
Calculates n-day rolling averages & minimum annual n-day averages, fits a probability distribution to minimum annual average flows, and calculates low flow statistics (nQy) for discharge data.
See https://vt-hydroinformatics.github.io/lfas.html#fit-to-pearson-type-iii-distribution for information on the probability fitting approach used to calculate nQy in this function.
}
\examples{
# Basic usage
## Get data
library(dplyr)
gauge_data=dataRetrieval::readNWISdv(siteNumbers="10171000", startDate="2000-01-01", parameterCd="00060") \%>\% dplyr::rename(discharge_cfs=X_00060_00003)# Parameter code 00060 = discharge in cfs.
### Also try wqTools::findSites() to find other gauge locations.

## Run the function
JR_1700S_7Q10=nQy(data=gauge_data, n=7, y=10, date_col="Date", q_col="discharge_cfs")
JR_1700S_7Q10

## Other nQy types
JR1700S_1Q10=nQy(data=gauge_data, n=1, y=10, date_col="Date", q_col="discharge_cfs")
JR1700S_1Q10

}
