% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permits_coords.R
\docType{data}
\name{permits_coords}
\alias{permits_coords}
\title{Utah DWQ permit coordinates}
\format{
Data frame
}
\usage{
permits_coords
}
\description{
Data frame containing permit IDs, names, and coordinates
}
\keyword{datasets}
