% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updatePermits.R
\name{updatePermits}
\alias{updatePermits}
\title{Update permit locations}
\usage{
updatePermits()
}
\value{
Exports a .rdata object of permit identifiers and spatial coordinates to the extdata folder of wasteloadR.
}
\description{
This function updates permit locations in the local install of wasteloadR. Queries permit locations from EPA ECHO. Primarily for use by package developers.
}
\examples{
updatePermits()
}
