% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wasteloadR.r
\name{wasteloadR}
\alias{wasteloadR}
\title{Run the wasteloadR Dashboard
Runs the wasteloadR dashboard app embedded in UT wasteloadR package.}
\usage{
wasteloadR()
}
\description{
Run the wasteloadR Dashboard
Runs the wasteloadR dashboard app embedded in UT wasteloadR package.
}
