% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wetCalc.R
\name{wetCalc}
\alias{wetCalc}
\title{Whole effluent toxicity (WET) test calculations for Utah River's model}
\usage{
wetCalc(
  critQ_cfs,
  effluentQ_cfs,
  width_ft,
  plume_width_15min_ft,
  plume_width_2500ft_ft
)
}
\arguments{
\item{critQ_cfs}{Upstream seasonal critical flow (typically 7Q10) in CFS}

\item{effluentQ_cfs}{Seasonal average effluent discharge in CFS}

\item{width_ft}{Stream channel width in feet (also used in mixZone())}

\item{plume_width_15min_ft}{Plume width at 15 min or 50\%: determined in mixing zone function (mixZone() output)}

\item{plume_width_2500ft_ft}{Plume width at 2500 ft: determined in mixing zone function (mixZone() output)}
}
\value{
TBD
}
\description{
Whole effluent toxicity (WET) test calculations for Utah River's model
}
