% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nQy.R
\name{nQy}
\alias{nQy}
\title{Generate low flow statistics (nQy) for wasteloads}
\usage{
nQy(
  data,
  n = 7,
  y = 10,
  date_col = "Date",
  q_col = "discharge_cfs",
  plot_fit = TRUE
)
}
\arguments{
\item{data}{Input data. Should be a dataframe of daily discharge values}

\item{n}{Numeric. Days for rolling time period averages. Default 7.}

\item{y}{Numeric. Recurrence interval in years}

\item{date_col}{Column name containing date values. Must be in standard "YYYY-MM-DD" date format.}

\item{q_col}{Column name containing discharge values. Any appropriate discharge units are OK. The result will be in the same units as the input.}
}
\value{
Returns a list of results including the calculated nQy statistic, a data frame of annual n-day flow minima and distribution information, and a plot to examine the fit.
}
\description{
Generate low flow statistics (nQy) for wasteloads
}
\examples{
# Basic usage
## Get data
library(dplyr)
gauge_data=dataRetrieval::readNWISdv(siteNumbers="10171000", startDate="2000-01-01", parameterCd="00060") \%>\% dplyr::rename(discharge_cfs=X_00060_00003, disch_code=X_00060_00003_cd)# Parameter code 00060 = discharge in cfs. see also ?readNWISuv() to read high frequency values
### Also try wqTools::findSites() to find other gauge locations.

## Run the function
JR_1700S_7Q10=nQy(data=gauge_data, n=7, y=10, date_col="Date", q_col="discharge_cfs")
JR_1700S_7Q10

## Other nQy types
JR1700S_1Q10=nQy(data=gauge_data, n=1, y=10, date_col="Date", q_col="discharge_cfs")
JR1700S_1Q10

}
